% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeP2024v1a.R
\name{ComputePValue}
\alias{ComputePValue}
\title{Compute p-values from information gains and return MDFS}
\usage{
ComputePValue(
  IG,
  dimensions,
  divisions,
  discretizations,
  contrast.mask = NULL,
  response.divisions = 1,
  df = NULL,
  ig.in.bits = TRUE,
  ig.doubled = FALSE,
  param.median = TRUE,
  n.outliers = NULL,
  max.outliers = NULL,
  level = 0.05
)
}
\arguments{
\item{IG}{max conditional information gains}

\item{dimensions}{number of dimensions}

\item{divisions}{number of divisions}

\item{discretizations}{number of discretizations (matters only for \code{dimensions=discretizations=1})}

\item{contrast.mask}{boolean mask on \code{IG} specifying which variables are contrast variables (or \code{NULL} if none, otherwise at least 3 variables must be marked). \code{NULL} is allowed only for \code{dimensions=discretizations=1}.}

\item{response.divisions}{number of response divisions (i.e. categories-1)}

\item{df}{vector of degrees of freedom for each variable (optional)}

\item{ig.in.bits}{\code{TRUE} if input is in binary log (as opposed to natural log)}

\item{ig.doubled}{\code{TRUE} if input is doubled (to follow the chi-squared distribution)}

\item{param.median}{\code{TRUE} if the null distribution parameter is estimated by the median of \code{IG} for contrast variables, rather than the mean value. This method is robust for outliers, although less precise.}

\item{n.outliers}{Specify \code{n.outliers} when the number of outliers (i.e. values of \code{IG} exceptionally close to zero) is known, e.g. to be zero.}

\item{max.outliers}{maximum possible number of outliers. Default value is 1/4 of the number of \code{IG} values.}

\item{level}{significance level for outlier detection. Change of the default value 0.05 is not recommended.}
}
\value{
A \code{\link{data.frame}} with class set to \code{MDFS}. Can be coerced back to \code{data.frame} using \code{\link{as.data.frame}}.

 The following columns are present:
 \itemize{
   \item \code{IG} -- information gains (input copy)
   \item \code{chi.squared.p.value} -- chi-squared p-values
   \item \code{p.value} -- theoretical p-values
 }

 Additionally the following \code{\link{attributes}} are set:
 \itemize{
  \item \code{run.params} -- run parameters
  \item \code{dist.param} -- distribution parameter
  \item \code{fdr.outliers} -- FILL ME WITH TEXT
  \item \code{n.outliers} -- FILL ME WITH TEXT
 }
}
\description{
Compute p-values from information gains and return MDFS
}
\examples{
ComputePValue(madelon$IG.2D, dimensions = 1, divisions = 1, discretizations=1)
}
