% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{find_fully_explained_vars_idx}
\alias{find_fully_explained_vars_idx}
\title{Find indexes of variables fully explained by a confounding factor}
\usage{
find_fully_explained_vars_idx(data, confounder)
}
\arguments{
\item{data}{\code{data.frame} of discrete variables}

\item{confounder}{discrete confounding variable}
}
\value{
a vector of indexes of variables in \code{data} which are fully explained by \code{confounder}
}
\description{
Find indexes of variables fully explained by a confounding factor
}
\examples{
\donttest{
data_bin<-as.data.frame(lapply(as.data.frame(madelon$data),
 function(x) 1.*(x>median(x)) ))
find_fully_explained_vars_idx(data_bin, data_bin[,2])
}
}
