% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscrMDFSwrapper.R
\name{MDFS.discrete}
\alias{MDFS.discrete}
\title{Run end-to-end MDFS for discretized data with same number of categories}
\usage{
MDFS.discrete(
  data,
  decision,
  n.contrast = max(ncol(data), 30),
  dimensions = 1,
  pc.xi = 0.25,
  p.adjust.method = "holm",
  level = 0.05,
  seed = NULL
)
}
\arguments{
\item{data}{input data.frame where columns are variables and rows are observations (all numeric)}

\item{decision}{decision variable as a boolean vector of length equal to number of observations}

\item{n.contrast}{number of constrast variables (defaults to max of 1/10 of variables number and 30)}

\item{dimensions}{number of dimensions}

\item{pc.xi}{parameter xi used to compute pseudocounts (the default is recommended not to be changed)}

\item{p.adjust.method}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"} is recommended for FDR, see Details)}

\item{level}{statistical significance level}

\item{seed}{seed for PRNG used during discretizations (\code{NULL} for random)}
}
\value{
A \code{\link{list}} with the following fields:
 \itemize{
   \item \code{contrast.indices} -- indices of variables chosen to build contrast variables
   \item \code{statistic} -- vector of statistic's values (IGs) for corresponding variables
   \item \code{p.value} -- vector of p-values for corresponding variables
   \item \code{adjusted.p.value} -- vector of adjusted p-values for corresponding variables
   \item \code{relevant.variables} -- vector of relevant variables indices
 }
}
\description{
Run end-to-end MDFS for discretized data with same number of categories
}
\details{
In case of FDR control it is recommended to use Benjamini-Hochberg-Yekutieli p-value adjustment
method (\code{"BY"} in \code{\link[stats]{p.adjust}}) due to unknown dependencies between tests.
}
\examples{
\donttest{
as.data.frame(lapply(as.data.frame(madelon$data), function(COL) 1.*(COL > median(COL)) ) )-> discr_data
MDFS.discrete(discr_data, madelon$decision, dimensions = 2)
}
}
