% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{opt.discretize}
\alias{opt.discretize}
\title{Discretize continuous variable in an optimal way according to p-value of mutual information with discrete y}
\usage{
opt.discretize(x, y, divisions = 1, range = 0.5, grid = 10, seed = NULL)
}
\arguments{
\item{x}{numeric vector, a continuous variable to be discretized}

\item{y}{numeric vector, discrete variable}

\item{divisions}{number of divisions of \code{x} (n_classes -1)}

\item{range}{discretization range}

\item{grid}{number of random tries per each class of discretized \code{x}}

\item{seed}{random seed for randomized discretization procedure}
}
\value{
a \code{list} with the following fields: 
 \itemize{
   \item \code{breaks.opt} -- division points of \code{x} in optimal discretization
   \item \code{pv.opt} -- chi-squared p-value of optimal discretization
   \item \code{x} -- discretized \code{x}
 }
}
\description{
Discretize continuous variable in an optimal way according to p-value of mutual information with discrete y
}
\examples{
\donttest{
x<- runif(1000)
opt.discretize(x, (x>0.5)*1.)
}
}
