% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{pack.confounders}
\alias{pack.confounders}
\title{Pack many confounders into a single value}
\usage{
pack.confounders(confounders, all.combinations = F)
}
\arguments{
\item{confounders}{object containing discrete variables, gets coerced to \code{data.frame}}

\item{all.combinations}{boolean stating whether to include all possible combinations in the packed confounder (\code{TRUE}) or only those appearing in data provided (\code{FALSE})}
}
\value{
a vector containing discrete variable, a packed confounder with one value per each combination of \code{confounders}
}
\description{
Pack many confounders into a single value
}
\examples{
\donttest{
discr_data<- as.data.frame(lapply(as.data.frame(madelon$data[,1:3]), 
				function(column) (column > median(column) )*1. ))
pack.confounders(discr_data, all.combinations=TRUE)
}
}
