% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMI.R
\name{pairwiseMIsimilarity}
\alias{pairwiseMIsimilarity}
\title{Compute similarity matrix of continuous data based on their mutual information after discretization}
\usage{
pairwiseMIsimilarity(
  data,
  divisions = 1,
  discretizations = 1,
  n.contrast = max(ncol(data), 30),
  p.adjust.method = "holm",
  level = 0.05,
  seed = NULL
)
}
\arguments{
\item{data}{input data.frame where columns are variables and rows are observations (all numeric, discrete variables with same number of categories)}

\item{divisions}{number of divisions}

\item{discretizations}{number of discretizations}

\item{n.contrast}{number of contrast variables}

\item{p.adjust.method}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"} is recommended for FDR, see Details)}

\item{level}{statistical significance level}

\item{seed}{random seed for randomized discretization procedure}
}
\value{
A \code{\link{list}} of symmetric matrices, with the following fields:
 \itemize{
   \item \code{S} -- similarity matrix of variables where \code{S[i,j]} is equal to \code{MI[i,j]} if \code{pv.adjusted[i,j]} is \code{<level} else 0
   \item \code{pv.adjusted} -- \code{pv.adjusted[i,j]} is adjusted p-value of mutual information \code{MI[i,j]}. 
   \item \code{pv} -- \code{pv[i,j]} is p-value of mutual information \code{MI[i,j]} 
   \item \code{MI} -- \code{MI[i,j]} is mutual information computed between variables \code{data[,i]} and \code{data[,j]}
 }
}
\description{
Compute similarity matrix of continuous data based on their mutual information after discretization
}
\details{
In case of FDR control it is recommended to use Benjamini-Hochberg-Yekutieli p-value adjustment
method (\code{"BY"} in \code{\link[stats]{p.adjust}}) due to unknown dependencies between tests.
}
\examples{
\donttest{
pairwiseMIsimilarity(madelon$data, 1)
}
}
