% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{pairwiseMIsimilarityDiscrete.confounded}
\alias{pairwiseMIsimilarityDiscrete.confounded}
\title{Compute similarity matrix of discrete variables with same number of categories based on their mutual information taking into the account confounding variables}
\usage{
pairwiseMIsimilarityDiscrete.confounded(
  data,
  confounders,
  divisions = 1,
  p.adjust.method = "holm",
  level = 0.05
)
}
\arguments{
\item{data}{input data.frame where columns are variables and rows are observations (all numeric, discrete variables with same number of categories)}

\item{confounders}{object containing discrete variables, gets coerced to \code{data.frame}}

\item{divisions}{number of classes of each variable (same for all variables in data)}

\item{p.adjust.method}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"} is recommended for FDR, see Details)}

\item{level}{statistical significance level}
}
\value{
A \code{\link{list}} of symmetric matrices, with the following fields:
 \itemize{
   \item \code{S} -- similarity matrix of variables where \code{S[i,j]} is equal to \code{MI[i,j]} if \code{pv.adjusted[i,j]} is \code{<level} else 0
   \item \code{pv.adjusted} -- \code{pv.adjusted[i,j]} is adjusted p-value of mutual information \code{MI[i,j]}
   \item \code{pv} -- \code{pv[i,j]} is p-value of mutual information \code{MI[i,j]} 
   \item \code{MI} -- \code{MI[i,j]} is mutual information computed between variables \code{data[,i]} and \code{data[,j]}
 }
}
\description{
Compute similarity matrix of discrete variables with same number of categories based on their mutual information taking into the account confounding variables
}
\details{
In case of FDR control it is recommended to use Benjamini-Hochberg-Yekutieli p-value adjustment
method (\code{"BY"} in \code{\link[stats]{p.adjust}}) due to unknown dependencies between tests.
If any variable in \code{data} is fully explained by any confounder, execution stops.
To find fully explained variables one can use \code{find_fully_explained_vars_idx} function.
}
\examples{
\donttest{
discr_data<- as.data.frame(lapply(as.data.frame(madelon$data),
 function(column) (column > median(column) )*1. ))
confounder<- 1.*(madelon$data[,1] > median(madelon$data[,1]))
pairwiseMIsimilarityDiscrete.confounded(discr_data[,-1], confounder)
}
}
