% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{pv.1D}
\alias{pv.1D}
\title{p-value of mutual information of one discretized variable x and discrete y as a function of breaks}
\usage{
pv.1D(x, y, breaks)
}
\arguments{
\item{x}{numeric vector, a continuous variable to be discretized}

\item{y}{numeric vector, discrete variable}

\item{breaks}{vector of division points of \code{x} to test}
}
\value{
a chi-squared p-value of the mutual information
}
\description{
p-value of mutual information of one discretized variable x and discrete y as a function of breaks
}
\examples{
\donttest{
x<- runif(1000)
pv.1D(x, (x > 0.5)*1., c(0.2,0.5,0.7) )
}
}
