% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMI.R
\name{symmetrizeMatrix}
\alias{symmetrizeMatrix}
\title{Symmetrize matrix by choosing between one of lower/upper half elements}
\usage{
symmetrizeMatrix(A, pairwiseSelectionFunction = pmin)
}
\arguments{
\item{A}{square matrix to be symmetrized}

\item{pairwiseSelectionFunction}{vectorized function for producing element of a symmetrized matrix based on lower/upper half elements}
}
\value{
A symmetric matrix
}
\description{
Symmetrize matrix by choosing between one of lower/upper half elements
}
\examples{
symmetrizeMatrix(as.matrix(madelon$data[1:30,1:30]))
}
